+++ python/dist-mcp/pyproject.toml
@@ -4,8 +4,8 @@ version = "0.6.0"
 description = "Alation Agent SDK with MCP support"
 dependencies = [
   "alation-ai-agent-sdk~=0.6.0",
-  "mcp[cli]>=1.2.0",
-  "fastMCP==2.2.6"
+  "mcp[cli]>=1.12.0",
+  "fastMCP==2.9.0"
 ]
 authors = [
   { name="Jagannath (Jags) Saragadam", email="jags.saragadam@alation.com"},
